feat: 클래스명 및 url 수정
