+++ src/main/java/subway/domain/line/controller/LineController.java
@@ -1,22 +1,18 @@
 package subway.domain.line.controller;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import subway.domain.line.domain.Line;
-import subway.domain.line.domain.ShortestPath;
+import org.springframework.web.bind.annotation.*;
+import subway.domain.line.dto.LineRequest;
 import subway.domain.line.dto.LineResponse;
+import subway.domain.line.entity.LineEntity;
 import subway.domain.line.service.LineService;
 import subway.global.common.ResultResponse;
 
-import java.util.List;
-import java.util.stream.Collectors;
+import java.net.URI;
 
-@Controller
-@RequestMapping("/line")
+@RestController
+@RequestMapping("/line-detail")
 public class LineController {
 
     private final LineService lineService;
@@ -25,24 +21,22 @@ public class LineController {
         this.lineService = lineService;
     }
 
-    @GetMapping
-    public ResponseEntity<ResultResponse> findAllLine() {
-        List<Line> lines = lineService.findAll();
-        List<LineResponse> lineResponses = lines.stream()
-                .map(LineResponse::of)
-                .collect(Collectors.toList());
-        return ResponseEntity.ok().body(new ResultResponse(200, "전체 노선도 조회 성공", lineResponses));
+    @PostMapping
+    public ResponseEntity<ResultResponse> createLine(@RequestBody final LineRequest lineRequest) {
+        LineEntity lineEntity = lineService.saveLine(lineRequest);
+        LineResponse lineResponse = LineResponse.of(lineEntity);
+        return ResponseEntity.created(URI.create("/line-detail/" + lineResponse.getId())).body(new ResultResponse(201, "노선 추가 성공", lineResponse));
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<ResultResponse> findLineById(@PathVariable final Long id) {
-        Line line = lineService.findById(id);
-        return ResponseEntity.ok().body(new ResultResponse(200, "단일 노선도 조회 성공", LineResponse.of(line)));
+    @PutMapping("/{id}")
+    public ResponseEntity<ResultResponse> updateLine(@PathVariable final Long id, @RequestBody final LineRequest lineUpdateRequest) {
+        lineService.updateLine(id, lineUpdateRequest);
+        return ResponseEntity.ok().body(new ResultResponse(200, "노선 업데이트 성공", id));
     }
 
-    @GetMapping("/path")
-    public ResponseEntity<ResultResponse> findPath(@RequestParam final Long startLineId, @RequestParam final Long endLineId) {
-        ShortestPath shortestPath = lineService.findShortestPath(startLineId, endLineId);
-        return ResponseEntity.ok().body(new ResultResponse(200, "최단 거리 조회 성공", shortestPath));
+    @DeleteMapping("/{id}")
+    public ResponseEntity<ResultResponse> deleteLine(@PathVariable final Long id) {
+        lineService.deleteLineById(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new ResultResponse(204, "노선 삭제 성공", id));
     }
 }
