+++ src/main/java/subway/domain/line/dto/LineResponse.java
@@ -1,50 +1,50 @@
 package subway.domain.line.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import subway.domain.line.domain.Line;
-import subway.domain.lineDetail.dto.LineDetailResponse;
-import subway.domain.lineDetail.entity.LineDetailEntity;
-import subway.domain.station.dto.StationResponse;
-import subway.domain.station.entity.StationEntity;
+import subway.domain.line.entity.LineEntity;
 
-import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Objects;
 
 public class LineResponse {
 
-    @JsonProperty("lineDetail")
-    private LineDetailResponse lineDetailResponse;
-    @JsonProperty("stations")
-    private List<StationResponse> stationResponses;
+    private Long id;
+    private String name;
+    private String color;
 
     private LineResponse() {
     }
 
-    public LineResponse(LineDetailResponse lineDetailResponse, final List<StationResponse> stationResponses) {
-        this.lineDetailResponse = lineDetailResponse;
-        this.stationResponses = stationResponses;
+    public LineResponse(final Long id, final String name, final String color) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
     }
 
-    public static LineResponse of(Line line) {
-        LineDetailEntity lineDetail = line.getLineDetail();
-        List<StationEntity> stations = line.getStations();
-
-        LineDetailResponse lineDetailResponse = new LineDetailResponse(lineDetail.getId(),
-                lineDetail.getName(),
-                lineDetail.getColor());
-
-        List<StationResponse> stationResponses = stations.stream().map(
-                stationEntity -> new StationResponse(stationEntity.getId(), stationEntity.getName())
-        ).collect(Collectors.toList());
-
-        return new LineResponse(lineDetailResponse, stationResponses);
+    public static LineResponse of(final LineEntity lineEntity) {
+        return new LineResponse(lineEntity.getId(), lineEntity.getName(), lineEntity.getColor());
     }
 
-    public LineDetailResponse getLineDetail() {
-        return lineDetailResponse;
+    public Long getId() {
+        return id;
     }
 
-    public List<StationResponse> getStations() {
-        return stationResponses;
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        LineResponse that = (LineResponse) o;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
     }
 }
