+++ src/main/java/subway/domain/path/domain/FareCalculator.java
@@ -1,4 +1,4 @@
-package subway.domain.line.domain;
+package subway.domain.path.domain;
 
 public enum FareCalculator {
