--- src/test/java/subway/integration/LineDetailIntegrationTest.java
@@ -1,121 +1,0 @@
-package subway.integration;
-
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import subway.domain.lineDetail.dto.LineDetailRequest;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@DisplayName("지하철 노선 관련 기능")
-public class LineDetailIntegrationTest extends IntegrationTest {
-
-    private LineDetailRequest lineDetailRequest1;
-    private LineDetailRequest lineDetailRequest2;
-
-    @Override
-    @BeforeEach
-    public void setUp() {
-        super.setUp();
-
-        lineDetailRequest1 = new LineDetailRequest("신분당선", "bg-red-600");
-        lineDetailRequest2 = new LineDetailRequest("구신분당선", "bg-red-600");
-    }
-
-    @DisplayName("지하철 노선을 생성한다.")
-    @Test
-    void createLine() {
-        // when
-        final ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(lineDetailRequest1)
-                .when().post("/line-detail")
-                .then().log().all().
-                extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
-    @DisplayName("기존에 존재하는 지하철 노선 이름으로 지하철 노선을 생성한다.")
-    @Test
-    void createLineWithDuplicateName() {
-        // given
-        RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(lineDetailRequest1)
-                .when().post("/line-detail")
-                .then().log().all().
-                extract();
-
-        // when
-        final ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(lineDetailRequest1)
-                .when().post("/line-detail")
-                .then().log().all().
-                extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("지하철 노선을 수정한다.")
-    @Test
-    void updateLine() {
-        // given
-        final ExtractableResponse<Response> createResponse = RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(lineDetailRequest1)
-                .when().post("/line-detail")
-                .then().log().all().
-                extract();
-
-        // when
-        final Long lineId = Long.parseLong(createResponse.header("Location").split("/")[2]);
-        final ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(lineDetailRequest2)
-                .when().put("/line-detail/{lineId}", lineId)
-                .then().log().all()
-                .extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    @DisplayName("지하철 노선을 제거한다.")
-    @Test
-    void deleteLine() {
-        // given
-        final ExtractableResponse<Response> createResponse = RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(lineDetailRequest1)
-                .when().post("/line-detail")
-                .then().log().all().
-                extract();
-
-        // when
-        final Long lineId = Long.parseLong(createResponse.header("Location").split("/")[2]);
-        final ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .when().delete("/line-detail/{lineId}", lineId)
-                .then().log().all()
-                .extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-}
