+++ src/test/java/subway/integration/LineIntegrationTest.java
@@ -3,118 +3,119 @@ package subway.integration;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.domain.line.domain.ShortestPath;
-import subway.domain.line.dto.LineResponse;
-import subway.domain.lineDetail.dto.LineDetailResponse;
-import subway.domain.station.dto.StationResponse;
-import subway.domain.station.entity.StationEntity;
-
-import java.util.List;
+import subway.domain.line.dto.LineRequest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
-@DisplayName("지하철 지도 관련 기능")
-class LineIntegrationTest extends IntegrationTest {
+@DisplayName("지하철 노선 관련 기능")
+public class LineIntegrationTest extends IntegrationTest {
 
-    @DisplayName("지하철 노선을 조회한다.")
-    @Test
-    void getLine() {
-        // when
-        final ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/line/{lineId}", 1L)
-                .then().log().all()
-                .extract();
+    private LineRequest lineRequest1;
+    private LineRequest lineRequest2;
 
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    @Override
+    @BeforeEach
+    public void setUp() {
+        super.setUp();
 
-        final LineResponse lineResponse = response.jsonPath().getObject("data.", LineResponse.class);
-
-        assertThat(lineResponse.getLineDetail()).isEqualTo(new LineDetailResponse(1L, "2호선", "초록색"));
-        assertThat(lineResponse.getStations()).containsExactly(
-                new StationResponse(1L, "신림역"),
-                new StationResponse(2L, "봉천역"),
-                new StationResponse(3L, "서울대입구역"),
-                new StationResponse(4L, "낙성대역"),
-                new StationResponse(5L, "사당역"),
-                new StationResponse(6L, "방배역"),
-                new StationResponse(7L, "서초역")
-        );
+        lineRequest1 = new LineRequest("신분당선", "bg-red-600");
+        lineRequest2 = new LineRequest("구신분당선", "bg-red-600");
     }
 
-    @DisplayName("지하철 노선 목록을 조회한다.")
+    @DisplayName("지하철 노선을 생성한다.")
     @Test
-    void getLines() {
+    void createLine() {
         // when
         final ExtractableResponse<Response> response = RestAssured
                 .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/line")
-                .then().log().all()
-                .extract();
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(lineRequest1)
+                .when().post("/line-detail")
+                .then().log().all().
+                extract();
 
         // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-
-
-        List<LineResponse> result = response.jsonPath().getList("data.", LineResponse.class);
-
-        assertAll(
-                () -> assertThat(result.get(0).getLineDetail()).isEqualTo(new LineDetailResponse(1L, "2호선", "초록색")),
-                () -> assertThat(result.get(0).getStations()).containsExactly(
-                        new StationResponse(1L, "신림역"),
-                        new StationResponse(2L, "봉천역"),
-                        new StationResponse(3L, "서울대입구역"),
-                        new StationResponse(4L, "낙성대역"),
-                        new StationResponse(5L, "사당역"),
-                        new StationResponse(6L, "방배역"),
-                        new StationResponse(7L, "서초역")
-                ),
-                () -> assertThat(result.get(1).getLineDetail()).isEqualTo(new LineDetailResponse(2L, "3호선", "파란색")),
-                () -> assertThat(result.get(1).getStations()).containsExactly(
-                        new StationResponse(8L, "교대역"),
-                        new StationResponse(9L, "강남역"),
-                        new StationResponse(4L, "낙성대역"),
-                        new StationResponse(10L, "역삼역"),
-                        new StationResponse(11L, "선릉역")
-                )
-        );
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
     }
 
+    @DisplayName("기존에 존재하는 지하철 노선 이름으로 지하철 노선을 생성한다.")
     @Test
-    void 최단거리를_조회한다() {
+    void createLineWithDuplicateName() {
+        // given
+        RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(lineRequest1)
+                .when().post("/line-detail")
+                .then().log().all().
+                extract();
+
         // when
         final ExtractableResponse<Response> response = RestAssured
                 .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/line/path?startLineId=8&endLineId=7")
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(lineRequest1)
+                .when().post("/line-detail")
+                .then().log().all().
+                extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
+
+    @DisplayName("지하철 노선을 수정한다.")
+    @Test
+    void updateLine() {
+        // given
+        final ExtractableResponse<Response> createResponse = RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(lineRequest1)
+                .when().post("/line-detail")
+                .then().log().all().
+                extract();
+
+        // when
+        final Long lineId = Long.parseLong(createResponse.header("Location").split("/")[2]);
+        final ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(lineRequest2)
+                .when().put("/line-detail/{lineId}", lineId)
                 .then().log().all()
                 .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
 
-        final ShortestPath shortestPath = response.jsonPath().getObject("data.", ShortestPath.class);
+    @DisplayName("지하철 노선을 제거한다.")
+    @Test
+    void deleteLine() {
+        // given
+        final ExtractableResponse<Response> createResponse = RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(lineRequest1)
+                .when().post("/line-detail")
+                .then().log().all().
+                extract();
 
-        Assertions.assertAll(
-                () -> assertThat(shortestPath.getPath()).containsExactly(
-                        new StationEntity(8L, "교대역"),
-                        new StationEntity(9L, "강남역"),
-                        new StationEntity(4L, "낙성대역"),
-                        new StationEntity(5L, "사당역"),
-                        new StationEntity(6L, "방배역"),
-                        new StationEntity(7L, "서초역")
-                ),
-                () -> assertThat(shortestPath.getDistance()).isEqualTo(35.0),
-                () -> assertThat(shortestPath.getFare()).isEqualTo(1_750)
-        );
+        // when
+        final Long lineId = Long.parseLong(createResponse.header("Location").split("/")[2]);
+        final ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .when().delete("/line-detail/{lineId}", lineId)
+                .then().log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
     }
 }
