+++ src/test/java/subway/section/entity/SectionsMapperTest.java
@@ -1,4 +1,4 @@
-package subway.line.domain;
+package subway.section.entity;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
